+++ docs/changelog.md
@@ -1,6 +1,6 @@
 
 
-**The best Elasticsearch Highlevel Rest  Client API-----[bboss](https://esdoc.bbossgroups.com/#/README)**   v6.0.3 发布。
+**The best Elasticsearch Highlevel Rest  Client API-----[bboss](https://esdoc.bbossgroups.com/#/README)**   v6.0.5 发布。
 
 https://esdoc.bbossgroups.com/#/quickstart
 
@@ -34,7 +34,7 @@ https://esdoc.bbossgroups.com/#/development
         <dependency>
             <groupId>com.bbossgroups.plugins</groupId>
             <artifactId>bboss-elasticsearch-rest-jdbc</artifactId>
-            <version>6.0.3</version>
+            <version>6.0.5</version>
         </dependency>
 ```
 
@@ -44,11 +44,24 @@ https://esdoc.bbossgroups.com/#/development
         <dependency>
             <groupId>com.bbossgroups.plugins</groupId>
             <artifactId>bboss-elasticsearch-spring-boot-starter</artifactId>
-            <version>6.0.3</version>
+            <version>6.0.5</version>
         </dependency>
 ```
 
+# v6.0.5 功能改进 
 
+1.修复bug：
+动态通过ESIndex设置索引名称和索引type不起作用bug
+数据同步中context动态设置索引名称和索引type不起作用bug
+
+2.@ESIndex注解改进：增加useBatchContextIndexName属性
+
+```java
+ /* 当ESIndex的name只是用当前时间作为后缀，那么useBatchContextIndexName可以设置为true，提升解析动态索引名称性能，否则保持默认值false
+ */
+//@ESIndex(name="demowithesindex-{agentStarttime,yyyy.MM.dd}",type="demowithesindex")
+@ESIndex(name="demowithesindex-{dateformat=yyyy.MM.dd}",useBatchContextIndexName = true)
+```
 
 # v6.0.3 功能改进 
 1.http组件改进: 在异常信息中包含服务请求完整url地址信息
@@ -56,6 +69,7 @@ https://esdoc.bbossgroups.com/#/development
 2.http proxy组件改进：如果http服务池没有配置health状态检查地址，启用被动的服务健康检查机制，在没有正常节点的情况下，返回异常节点，如果操作成功则将异常节点标注为正常节点
 
 3.http组件改造：增加automaticRetriesDisabled开关，没有指定重试机制的情况下，如果automaticRetriesDisabled为false，在通讯则失败时自动重试3次，否则不重试
+
 # v6.0.2 功能改进 
 kafka2x-elasticsearch数据同步改进：kafka2x改进，提升同步性能